+++ check-creds/index.js
@@ -1,0 +1,29 @@
+/**
+ * Credential check
+ * - aws-lite requires credentials to initialize
+ * - Architect needs credentials for some things (e.g. Deploy), but also has a variety of offline workflows that interface with AWS service API emulators (e.g. Sandbox)
+ * - Thus, sometimes it's ok to use dummy creds, but sometimes we need to halt (via this util)
+ */
+module.exports = function checkAwsCredentials (params, callback) {
+  // eslint-disable-next-line
+  let awsLite = require('@aws-lite/client')
+  let { inventory } = params
+
+  let promise
+  if (!callback) {
+    promise = new Promise((res, rej) => {
+      callback = (err, result) => err ? rej(err) : res(result)
+    })
+  }
+
+  let errMsg = 'Valid AWS credentials needed to continue; missing or invalid credentials'
+  awsLite({
+    autoloadPlugins: false,
+    profile: inventory.inv?.aws?.profile, // aws-lite falls back to AWS_PROFILE or 'default' if undefined
+    region: 'us-west-1',
+  })
+    .then(() => callback())
+    .catch(() => callback(Error(errMsg)))
+
+  return promise
+}