+++ test/banner/cred-check-test.js
@@ -1,5 +1,5 @@
 let test = require('tape')
-let credCheck = require('../../banner/cred-check')
+let checkCreds = require('../../check-creds')
 
 function reset (t) {
   let envVars = [
@@ -21,33 +21,54 @@ let inventory = { inv: { aws: {} } }
 
 test('Set up env', t => {
   t.plan(1)
-  t.ok(credCheck, 'Found credCheck')
+  t.ok(checkCreds, 'Found checkCreds')
 })
 
-test('Credential check is disabled', t => {
+test('Credential checks (async)', async t => {
   t.plan(2)
-  let err = credCheck({ checkCreds: false, inventory })
-  t.notOk(err, 'No credential loading error reported')
-  t.notOk(process.env.ARC_AWS_CREDS, 'Did not mutate ARC_AWS_CREDS')
-  reset(t)
-})
-
-test('Credential checks', t => {
-  t.plan(3)
-  let err
 
   // Count on aws-lite finding creds (via env)
   process.env.AWS_ACCESS_KEY_ID = 'yo'
   process.env.AWS_SECRET_ACCESS_KEY = 'yo'
-  err = credCheck({ inventory })
-  t.notOk(err, 'No credential loading error reported')
-  t.notOk(process.env.ARC_AWS_CREDS, 'Did not mutate ARC_AWS_CREDS')
+  try {
+    await checkCreds({ inventory })
+    t.pass('No credential loading error reported')
+  }
+  catch (err) {
+    t.fail(err)
+  }
 
   // Fail a cred check
   reset(t)
   process.env.AWS_PROFILE = 'random_profile_name_that_does_not_exist'
-  err = credCheck({ inventory, needsValidCreds: true })
-  t.ok(err, 'Reported credential loading error')
-  console.log(err)
+  try {
+    await checkCreds({ inventory })
+    t.fail('Should have errored')
+  }
+  catch (err) {
+    t.ok(err, 'Reported credential loading error')
+  }
   reset(t)
+})
+
+test('Credential checks (callback)', t => {
+  t.plan(2)
+
+  // Count on aws-lite finding creds (via env)
+  process.env.AWS_ACCESS_KEY_ID = 'yo'
+  process.env.AWS_SECRET_ACCESS_KEY = 'yo'
+  checkCreds({ inventory }, err => {
+    reset(t)
+    if (err) t.fail(err)
+    else t.pass('No credential loading error reported')
+  })
+
+  // Fail a cred check
+  reset(t)
+  process.env.AWS_PROFILE = 'random_profile_name_that_does_not_exist'
+  checkCreds({ inventory }, err => {
+    reset(t)
+    if (err) t.ok(err, 'Reported credential loading error')
+    else t.fail('Should have errored')
+  })
 })